requirements.txt
flask
supabase
python-dotenv
scikit-learn
joblib
flask-bcrypt
bcrypt
